Squirrel large file support

+++ src/Update/Program.cs
@@ -132,7 +132,7 @@ namespace Squirrel.Update
                     break;
 #endif
                 case UpdateAction.Releasify:
-                    Releasify(opt.target, opt.releaseDir, opt.packagesDir, opt.bootstrapperExe, opt.backgroundGif, opt.signingParameters, opt.baseUrl, opt.setupIcon, !opt.noMsi, opt.packageAs64Bit, opt.frameworkVersion, !opt.noDelta);
+                    Releasify(opt.target, opt.releaseDir, opt.packagesDir, opt.bootstrapperExe, opt.backgroundGif, opt.signingParameters, opt.baseUrl, opt.setupIcon, !opt.noMsi, opt.packageAs64Bit, opt.frameworkVersion, !opt.noDelta, opt.noSetups);
                     break;
                 }
             }
@@ -296,7 +296,7 @@ namespace Squirrel.Update
             }
         }
 
-        public void Releasify(string package, string targetDir = null, string packagesDir = null, string bootstrapperExe = null, string backgroundGif = null, string signingOpts = null, string baseUrl = null, string setupIcon = null, bool generateMsi = true, bool packageAs64Bit = false, string frameworkVersion = null, bool generateDeltas = true)
+        public void Releasify(string package, string targetDir = null, string packagesDir = null, string bootstrapperExe = null, string backgroundGif = null, string signingOpts = null, string baseUrl = null, string setupIcon = null, bool generateMsi = true, bool packageAs64Bit = false, string frameworkVersion = null, bool generateDeltas = true, bool noSetups = false)
         {
             ensureConsole();
 
@@ -393,39 +393,42 @@ namespace Squirrel.Update
 
             ReleaseEntry.WriteReleaseFile(releaseEntries, releaseFilePath);
 
-            var targetSetupExe = Path.Combine(di.FullName, "Setup.exe");
-            var newestFullRelease = releaseEntries.MaxBy(x => x.Version).Where(x => !x.IsDelta).First();
+			// WriteZipToSetup utility cannot create setups bigger than 2GB.
+			if (!noSetups) {
+				var targetSetupExe = Path.Combine(di.FullName, "Setup.exe");
+				var newestFullRelease = releaseEntries.MaxBy(x => x.Version).Where(x => !x.IsDelta).First();
 
-            File.Copy(bootstrapperExe, targetSetupExe, true);
-            var zipPath = createSetupEmbeddedZip(Path.Combine(di.FullName, newestFullRelease.Filename), di.FullName, backgroundGif, signingOpts, setupIcon).Result;
+				File.Copy(bootstrapperExe, targetSetupExe, true);
+				var zipPath = createSetupEmbeddedZip(Path.Combine(di.FullName, newestFullRelease.Filename), di.FullName, backgroundGif, signingOpts, setupIcon).Result;
 
-            var writeZipToSetup = Utility.FindHelperExecutable("WriteZipToSetup.exe");
+				var writeZipToSetup = Utility.FindHelperExecutable("WriteZipToSetup.exe");
 
-            try {
-                var arguments = String.Format("\"{0}\" \"{1}\" \"--set-required-framework\" \"{2}\"", targetSetupExe, zipPath, frameworkVersion);
-                var result = Utility.InvokeProcessAsync(writeZipToSetup, arguments, CancellationToken.None).Result;
-                if (result.Item1 != 0) throw new Exception("Failed to write Zip to Setup.exe!\n\n" + result.Item2);
-            } catch (Exception ex) {
-                this.Log().ErrorException("Failed to update Setup.exe with new Zip file", ex);
-            } finally {
-                File.Delete(zipPath);
-            }
+				try {
+				    var arguments = String.Format("\"{0}\" \"{1}\" \"--set-required-framework\" \"{2}\"", targetSetupExe, zipPath, frameworkVersion);
+				    var result = Utility.InvokeProcessAsync(writeZipToSetup, arguments, CancellationToken.None).Result;
+				    if (result.Item1 != 0) throw new Exception("Failed to write Zip to Setup.exe!\n\n" + result.Item2);
+				} catch (Exception ex) {
+				    this.Log().ErrorException("Failed to update Setup.exe with new Zip file", ex);
+				} finally {
+				    File.Delete(zipPath);
+				}
 
-            Utility.Retry(() =>
-                setPEVersionInfoAndIcon(targetSetupExe, new ZipPackage(package), setupIcon).Wait());
+				Utility.Retry(() =>
+				    setPEVersionInfoAndIcon(targetSetupExe, new ZipPackage(package), setupIcon).Wait());
 
-            if (signingOpts != null) {
-                signPEFile(targetSetupExe, signingOpts).Wait();
-            }
+				if (signingOpts != null) {
+				    signPEFile(targetSetupExe, signingOpts).Wait();
+				}
 
-            if (generateMsi) {
-                createMsiPackage(targetSetupExe, new ZipPackage(package), packageAs64Bit).Wait();
+				if (generateMsi) {
+				    createMsiPackage(targetSetupExe, new ZipPackage(package), packageAs64Bit).Wait();
 
-                if (signingOpts != null) {
-                    signPEFile(targetSetupExe.Replace(".exe", ".msi"), signingOpts).Wait();
-                }
-            }
-        }
+				    if (signingOpts != null) {
+				        signPEFile(targetSetupExe.Replace(".exe", ".msi"), signingOpts).Wait();
+				    }
+				}
+			}
+		}
 
         public void Shortcut(string exeName, string shortcutArgs, string processStartArgs, string icon, bool onlyUpdate)
         {
